+++ src/operations/run/RunOperations.ts
@@ -2,11 +2,12 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+import { OperationUtils } from "../OperationUtils";
 import { OperationsBase } from "../../base/OperationsBase";
 import { PreferReturn } from "../../base/interfaces/CommonInterfaces";
-import { MinimalRun, ResponseFromGetRun, ResponseFromGetRunList, ResponseFromGetRunListMinimal, RunDetails } from "../../base/interfaces/apiEntities/RunInterfaces";
-import { OperationOptions } from "../OperationOptions";
-import { ParamsToDeleteRun, ParamsToGetRun, ParamsToGetRunList } from "./RunOperationParams";
+import type { MinimalRun, ResponseFromGetRun, ResponseFromGetRunList, ResponseFromGetRunListMinimal, RunDetails } from "../../base/interfaces/apiEntities/RunInterfaces";
+import type { OperationOptions } from "../OperationOptions";
+import type { ParamsToDeleteRun, ParamsToGetRun, ParamsToGetRunList } from "./RunOperationParams";
 
 export class RunOperations<TOptions extends OperationOptions> extends OperationsBase<TOptions> {
   constructor(
@@ -24,7 +25,9 @@ export class RunOperations<TOptions extends OperationOptions> extends Operations
    * @returns {Promise<MinimalRun[]>} minimal Run list. See {@link MinimalRun}.
    */
   public async getMinimalList(params: ParamsToGetRunList): Promise<MinimalRun[]> {
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendGetRequest<ResponseFromGetRunListMinimal>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getRunListUrl({ urlParams: params.urlParams }),
       preferReturn: PreferReturn.Representation,
@@ -41,7 +44,9 @@ export class RunOperations<TOptions extends OperationOptions> extends Operations
    * @returns {Promise<RunDetails[]>} array of Run details. See {@link RunDetails}.
    */
   public async getRepresentationList(params: ParamsToGetRunList): Promise<RunDetails[]> {
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendGetRequest<ResponseFromGetRunList>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getRunListUrl({ urlParams: params.urlParams }),
       preferReturn: PreferReturn.Representation,
@@ -58,7 +63,9 @@ export class RunOperations<TOptions extends OperationOptions> extends Operations
    */
   public async getSingle(params: ParamsToGetRun): Promise<RunDetails> {
     const { accessToken, runId } = params;
+    OperationUtils.ensureAccessTokenProvided(accessToken, this._options.accessTokenCallback);
     const response = await this.sendGetRequest<ResponseFromGetRun>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getSingleRunUrl({ runId }),
     });
@@ -72,7 +79,9 @@ export class RunOperations<TOptions extends OperationOptions> extends Operations
    * @returns {Promise<void>}.
    */
   public async delete(params: ParamsToDeleteRun): Promise<void> {
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     await this.sendDeleteRequest<void>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.deleteRunUrl(params),
     });
